Grant access to all objects till scheduler is started
An unprivileged task cannot access any kernel object other than itself
or the objects it has been explicitly granted access. These access
permissions are associated with the task. At runtime, the calling task's
access permissions are used for authorization when an unprivileged task
makes a system call.

This leaves the case when a system call is made before the scheduler is
started - no task is running and therefore, we cannot use any task's
access permissions for authorization. This commit updates the code to
authorize access to all the kernel objects before the scheduler is
started. The rationale is that the processor runs in privileged mode
before the scheduler is started.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>