Add traceQUEUE_SEND{_FROM_ISR,}_EXT and traceQUEUE_RESET hooks.
The current hooks were not effective at allowing a tracer to track the
number of items in a queue, since it could not differentiate between a
queueOVERWRITE, queueSEND_TO_BACK, and queueSEND_TO_FRONT send, and
could not (efficiently) trace a queue reset.

For sends, this adds extended tracing macros that, if not implemented,
fall back on the original tracing macros for backwards compatibility,
and introduces a new traceQUEUE_RESET hook.

Discussed here: https://forums.freertos.org/t/queue-tracing/20054/4