Split traceTASK_NOTIFY_WAIT into _EXT and _FAILED hooks.
The old traceTASK_NOTIFY_WAIT hook was always called (no matter the
success or failure of the action) and did not hygienically expose
enough information for the tracer to determine the state of the task
notification after it has been taken.

The new traceTASK_NOTIFY_WAIT_EXT hook is called only if
the notification was taken successfully and hygienically expose
ulBitsToClearOnExit.

The new traceTASK_NOTIFY_WAIT_FAILED hook is called if the notification
could not be taken.

This matches how the tracing hooks for all other APIs that attempt to
receive/take a resource and my block function: First, if the task blocks,
a BLOCK or BLOCKING trace hook is called. Then, either a normal or
FAILED trace hook is called, indicating if the operation timed out or
succeeded.

Both hooks fall back on the old traceTASK_NOTIFY_WAIT, preserving its
functionality for backwards compatibility.

Not that there is a very slight breaking change in this commit:
traceTASK_NOTIFY_WAIT is now called after the out-var pulNotificationValue
is set. Because this pointer was in an unknown/user-set state when the
tracing hook was previously called, it is highly unlikely that there are
any tracers that rely on this.