fix(freertos/smp): Bring back prvCheckForRunStateChange()
When granular locks are enabled, it may so happen that 2 tasks are
contesting for the kernel lock wherein 1 task gets the lock and puts the
other task in to suspend state. Then the other tasks gets the lock and
can put itself back into ready or delayed state, therefore the task
state becomes corrupt/unexpected. Hence, prvCheckForRunStateChange() is
needed when we suspend the kernel.