refactor: change methods ENTER|EXIT critical
The read and write from BaseType_t, can be do in a atomic context, and
dont need a mutex contol.
Change this methods for access the critical regiao only if is
neccessary. For make this, create a macro in portable.h that apoint to
tasENTER|EXIT_CRITIAL() function. Now, if port is guarantee that access
BaseType_t in atomic context, a empty macro can be define in
portmacro.h, like this:

Signed-off-by: guilherme giacomo simoes <trintaeoitogc@gmail.com>