[Fix] Reset correct priority on disinherit timeout
This commit corrects vTaskPriorityDisinheritAfterTimeout
to reset the previously inherited priority when the task
disiheriting timeout was the only task at that priority.
Without this change the ready list for the inherited
priority will remain set when no task is ready at that priority.
This can have consequences later as the ready priority flags are
assumed to be accurate.